Refactor Video state and functions into VideoPage.js component
