+++ src/App.js
@@ -1,4 +1,4 @@
-import { Button, Container, Stack } from "react-bootstrap";
+import { Container } from "react-bootstrap";
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 import Header from "./components/Header";
 import HomePage from "./pages/HomePage";
